# docs/openapi.yaml
openapi: 3.1.0
info:
  title: hodei-pipelines-server
  description: Hodei Pipelines Server - Monolithic Modular Architecture
  contact:
    name: Hodei Team
  license:
    name: MIT
    identifier: MIT
  version: 0.16.0
paths:
...
  /api/v1/worker-pools:
    get:
      tags:
      - worker-pools
      summary: Get all resource pools
      operationId: list_pools_handler
      responses:
        '200':
          description: List of resource pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourcePoolResponseDto'
        '500':
          description: Internal server error
    post:
      tags:
      - worker-pools
      summary: Create a new resource pool
      operationId: create_pool_handler
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePoolRequestDto'
        required: true
      responses:
        '200':
          description: Resource pool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePoolResponseDto'
        '400':
          description: Bad request
        '409':
          description: Pool already exists
        '500':
          description: Internal server error
  /api/v1/worker-pools/{pool_id}:
    get:
      tags:
      - worker-pools
      summary: Get a specific resource pool
      operationId: get_pool_handler
      parameters:
      - name: pool_id
        in: path
        description: Pool ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Resource pool details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePoolResponseDto'
        '404':
          description: Pool not found
        '500':
          description: Internal server error
    put:
      tags:
      - worker-pools
      summary: Update a resource pool (PUT - full update)
      operationId: update_pool_put_handler
      parameters:
      - name: pool_id
        in: path
        description: Pool ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePoolRequestDto'
        required: true
      responses:
        '200':
          description: Resource pool updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePoolResponseDto'
        '400':
          description: Bad request
        '404':
          description: Pool not found
        '500':
          description: Internal server error
    delete:
      tags:
      - worker-pools
      summary: Delete a resource pool
      operationId: delete_pool_handler
      parameters:
      - name: pool_id
        in: path
        description: Pool ID
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Resource pool deleted
        '404':
          description: Pool not found
        '500':
          description: Internal server error
    patch:
      tags:
      - worker-pools
      summary: Update a resource pool (PATCH - partial update) - US-10.4
      operationId: update_pool_patch_handler
      parameters:
      - name: pool_id
        in: path
        description: Pool ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePoolRequestDto'
        required: true
      responses:
        '200':
          description: Resource pool updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePoolResponseDto'
        '400':
          description: Bad request
        '404':
          description: Pool not found
        '500':
          description: Internal server error
  /api/v1/worker-pools/{pool_id}/status:
    get:
      tags:
      - worker-pools
      summary: Get resource pool status
      operationId: get_pool_status_handler
      parameters:
      - name: pool_id
        in: path
        description: Pool ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Resource pool status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcePoolStatusDto'
        '404':
          description: Pool status not found
        '500':
          description: Internal server error
components:
  schemas:
...
    CreatePoolRequestDto:
      type: object
      required:
      - poolType
      - name
      - providerName
      - minSize
      - maxSize
      - defaultResources
      properties:
        defaultResources:
          $ref: '#/components/schemas/ResourceQuotaDto'
        maxSize:
          type: integer
          format: int32
          minimum: 0
        minSize:
          type: integer
          format: int32
          minimum: 0
        name:
          type: string
        poolType:
          $ref: '#/components/schemas/ResourcePoolTypeDto'
        providerName:
          type: string
        tags:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
...
    ResourcePoolConfigDto:
      type: object
      required:
      - poolType
      - name
      - providerName
      - minSize
      - maxSize
      - defaultResources
      - tags
      properties:
        defaultResources:
          $ref: '#/components/schemas/ResourceQuotaDto'
        maxSize:
          type: integer
          format: int32
          minimum: 0
        minSize:
          type: integer
          format: int32
          minimum: 0
        name:
          type: string
        poolType:
          $ref: '#/components/schemas/ResourcePoolTypeDto'
        providerName:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
          propertyNames:
            type: string
    ResourcePoolResponseDto:
      type: object
      required:
      - id
      - config
      properties:
        config:
          $ref: '#/components/schemas/ResourcePoolConfigDto'
        id:
          type: string
    ResourcePoolStatusDto:
      type: object
      required:
      - name
      - poolType
      - totalCapacity
      - availableCapacity
      - activeWorkers
      - pendingRequests
      properties:
        activeWorkers:
          type: integer
          format: int32
          minimum: 0
        availableCapacity:
          type: integer
          format: int32
          minimum: 0
        name:
          type: string
        pendingRequests:
          type: integer
          format: int32
          minimum: 0
        poolType:
          $ref: '#/components/schemas/ResourcePoolTypeDto'
        totalCapacity:
          type: integer
          format: int32
          minimum: 0
    ResourcePoolTypeDto:
      type: string
      enum:
      - Docker
      - Kubernetes
      - Cloud
      - Static
    ResourceQuotaDto:
      type: object
      required:
      - cpu_m
      - memory_mb
      properties:
        cpu_m:
          type: integer
          format: int64
          minimum: 0
        gpu:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        memory_mb:
          type: integer
          format: int64
          minimum: 0
...
    UpdatePoolRequestDto:
      type: object
      properties:
        maxSize:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        minSize:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        name:
          type:
          - string
          - 'null'
        tags:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
...